% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exacerb}
\alias{exacerb}
\title{MS data}
\format{
An object of class \code{data.frame} with 765 rows and 14 columns.
}
\source{
Thal
}
\usage{
exacerb
}
\description{
MS data
}
\references{
Venab
}
\keyword{datasets}
