% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qif.r
\name{print.summary.qif}
\alias{print.summary.qif}
\title{Print the Summary of qif Object}
\usage{
\method{print}{summary.qif}(x, digits = NULL, quote = FALSE, prefix = "", ...)
}
\arguments{
\item{x}{the \code{qif} model object.}

\item{digits}{number of digits to print.}

\item{quote}{logical, indicating whether or not strings should be printed with
surrounding quotes.}

\item{prefix}{string, only \code{""} is implemented.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
The invisible object from the arguments.

The invisible object from the arguments.
}
\description{
Prints the summary of a \code{qif} object.
}
\seealso{
\code{\link[base]{summary}} \code{\link[qif]{qif}}
}
\author{
Zhichang Jiang, Alberta Health Services, and Peter X.K. Song, University
of Michigan.
}
