% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qif.r
\name{summary.qif}
\alias{summary.qif}
\title{Summary of a qif Object}
\usage{
\method{summary}{qif}(object, correlation = TRUE, ...)
}
\arguments{
\item{object}{an object for which a summary is desired.}

\item{correlation}{binary, include correlation.}
}
\value{
The \code{summary.qif} object.
}
\description{
Procuce a summary of a \code{qif} object.
}
\seealso{
\code{\link[qif]{qif}}
}
\author{
Zhichang Jiang, Alberta Health Services, and Peter X.K. Song, University
of Michigan.
}
